[CHORE] Moved SwaggerConfig
