+++ src/main/resources/application-sample.yml
@@ -9,3 +9,16 @@ spring:
     user:
       name:
       password:
+  cloud:
+    gcp:
+      credentials:
+        location: classpath:omoji-server-account.json
+  gcp:
+    config:
+      file: classpath:omoji-server-account.json
+    project:
+      id: omoji-server
+    bucket:
+      id: omoji-bucket
+    dir:
+      name: img
